Make company tooltip common faces consistent
I'm not sure if this is an intentional decision or potentially a bug. But normally (as found in most other modern editors), users would expect the tooltip common faces to be highlighted whether being selected or not. I find the behaviour to disable the common highlight when the candidate is selected to be inconsistent, and at times, confusing and less expressive.