Use plural in package name and add matching library
Every package should provide a library whose name matches that of the
package.  This allows tools, such as those used by Melpa, to extract
metadata.

At the same time every library named "*-theme.el" must itself
implement a theme because `load-theme' unfortunately assumes that
every file on `custom-theme-load-path' with such a name actually
provides a theme.  If it encounters files that to not provide a theme
but are never-the-less named "*-theme.el", then it falsely offers
those as themes, and trying to activate them will result in errors.

Furthermore, since this package provides multiple themes, it makes
sense to give it a name that suggests so.

Unfortunately package.el does not (yet?) support informing the user
about package renames or even automatically updating to the new name,
so some users will not notice this rename for a while.  Add a note
at the beginning of the readme, in the hope that will help at least
some users.